+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 0.0-18 (2019-06-11 / a27f7fc)
 
 ## Added